moves validator to core package and makes it pkg private
